move call to cms api to another file
